+++ core/chains/evm/types/types.go
@@ -85,6 +85,7 @@ type ChainCfg struct {
 	EvmHeadTrackerSamplingInterval                 *models.Duration
 	EvmLogBackfillBatchSize                        null.Int
 	EvmLogPollInterval                             *models.Duration
+	EvmLogKeepBlocksDepth                          null.Int
 	EvmMaxGasPriceWei                              *utils.Big
 	EvmNonceAutoSync                               null.Bool
 	EvmUseForwarders                               null.Bool