+++ contracts/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-...
+- Migrated and improved `AuthorizedReceiverInterface` and `AuthorizedReceiver` from 0.7.0
+- Added `OCR2DR` interfaces and contracts (initial version for PoC)
 
 ## 0.5.1 - 2022-09-27